Adjust compilation avoidance tests structure
