+++ subprojects/plugins/src/integTest/groovy/org/gradle/java/compile/AbstractJavaGroovyCompileAvoidanceIntegrationSpec.groovy
@@ -17,8 +17,6 @@
 package org.gradle.java.compile
 
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
-import org.gradle.language.fixtures.HelperProcessorFixture
-import org.junit.Assume
 import spock.lang.Issue
 
 abstract class AbstractJavaGroovyCompileAvoidanceIntegrationSpec extends AbstractIntegrationSpec {
@@ -281,7 +279,7 @@ public class ToolImpl {
         skipped ":b:${language.compileTaskName}"
     }
 
-    def "doesn't recompile when implementation class code changes for"() {
+    def "doesn't recompile when implementation class code changes"() {
         given:
         buildFile << """
             project(':b') {
@@ -523,99 +521,6 @@ public class ToolImpl {
         executedAndNotSkipped ":b:${language.compileTaskName}"
     }
 
-    def "doesn't recompile when private inner class changes"() {
-        given:
-        // Groovy doesn't produce real private inner classes - the generated bytecode has no ACC_PRIVATE
-        Assume.assumeTrue(language == Language.JAVA)
-        buildFile << """
-            project(':b') {
-                dependencies {
-                    compile project(':a')
-                }
-            }
-        """
-        def sourceFile = file("a/src/main/${language.name}/ToolImpl.${language.name}")
-        sourceFile << """
-            public class ToolImpl {
-                private class Thing { }
-            }
-        """
-        file("b/src/main/${language.name}/Main.${language.name}") << """
-            public class Main { ToolImpl t = new ToolImpl(); }
-        """
-
-        when:
-        succeeds ":b:${language.compileTaskName}"
-
-        then:
-        executedAndNotSkipped ":a:${language.compileTaskName}"
-        executedAndNotSkipped ":b:${language.compileTaskName}"
-
-        when:
-        // non-ABI change of inner class
-        sourceFile.text = """
-            public class ToolImpl {
-                private class Thing {
-                    public long v;
-                }
-            }
-"""
-
-        then:
-        succeeds ":b:${language.compileTaskName}"
-        executedAndNotSkipped ":a:${language.compileTaskName}"
-        skipped ":b:${language.compileTaskName}"
-
-        when:
-        // Remove inner class
-        sourceFile.text = """
-            public class ToolImpl {
-            }
-"""
-
-        then:
-        succeeds ":b:${language.compileTaskName}"
-        executedAndNotSkipped ":a:${language.compileTaskName}"
-        skipped ":b:${language.compileTaskName}"
-
-        when:
-        // Anonymous class
-        sourceFile.text = """
-            public class ToolImpl {
-                private Object r = new Runnable() { public void run() { } };
-            }
-"""
-
-        then:
-        succeeds ":b:${language.compileTaskName}"
-        executedAndNotSkipped ":a:${language.compileTaskName}"
-        skipped ":b:${language.compileTaskName}"
-
-        when:
-        // Add classes
-        sourceFile.text = """
-            public class ToolImpl {
-                private Object r = new Runnable() {
-                    public void run() {
-                        class LocalThing { }
-                    }
-                };
-                private static class C1 {
-                }
-                private class C2 {
-                    public void go() {
-                        Object r2 = new Runnable() { public void run() { } };
-                    }
-                }
-            }
-"""
-
-        then:
-        succeeds ":b:${language.compileTaskName}"
-        executedAndNotSkipped ":a:${language.compileTaskName}"
-        skipped ":b:${language.compileTaskName}"
-    }
-
     def "doesn't recompile when implementation resource is changed in various ways"() {
         given:
         buildFile << """
@@ -708,180 +613,6 @@ public class ToolImpl {
         succeeds ":b:${language.compileTaskName}"
         skipped ":a:${language.compileTaskName}"
         skipped ":b:${language.compileTaskName}"
-    }
-
-    def "recompiles source when annotation processor implementation on annotation processor classpath changes"() {
-        given:
-        Assume.assumeTrue(language == Language.JAVA)
-        settingsFile << "include 'c'"
-
-        buildFile << """
-            project(':b') {
-                dependencies {
-                    compile project(':a')
-                }
-            }
-            project(':c') {
-                configurations {
-                    processor
-                }
-                dependencies {
-                    compile project(':a')
-                    processor project(':b')
-                }
-                ${language.compileTaskName}.options.annotationProcessorPath = configurations.processor
-                task run(type: JavaExec) {
-                    main = 'TestApp'
-                    classpath = sourceSets.main.runtimeClasspath
-                }
-            }
-        """
-
-        def fixture = new HelperProcessorFixture()
-
-        // The annotation
-        fixture.writeApiTo(file("a"))
-
-        // The processor and library
-        fixture.writeSupportLibraryTo(file("b"))
-        fixture.writeAnnotationProcessorTo(file("b"))
-
-        // The class that is the target of the processor
-        file("c/src/main/${language.name}/TestApp.${language.name}") << '''
-            @Helper
-            class TestApp {
-                public static void main(String[] args) {
-                    System.out.println(new TestAppHelper().getValue()); // generated class
-                }
-            }
-'''
-
-        when:
-        run(':c:run')
-
-        then:
-        executedAndNotSkipped(":a:${language.compileTaskName}")
-        executedAndNotSkipped(":b:${language.compileTaskName}")
-        executedAndNotSkipped(":c:${language.compileTaskName}")
-        outputContains('greetings')
-
-        when:
-        run(':c:run')
-
-        then:
-        skipped(":a:${language.compileTaskName}")
-        skipped(":b:${language.compileTaskName}")
-        skipped(":c:${language.compileTaskName}")
-        outputContains('greetings')
-
-        when:
-        // Update the library class
-        fixture.message = 'hello'
-        fixture.writeSupportLibraryTo(file("b"))
-
-        run(':c:run')
-
-        then:
-        skipped(":a:${language.compileTaskName}")
-        executedAndNotSkipped(":b:${language.compileTaskName}")
-        executedAndNotSkipped(":c:${language.compileTaskName}")
-        outputContains('hello')
-
-        when:
-        run(':c:run')
-
-        then:
-        skipped(":a:${language.compileTaskName}")
-        skipped(":b:${language.compileTaskName}")
-        skipped(":c:${language.compileTaskName}")
-        outputContains('hello')
-
-        when:
-        // Update the processor class
-        fixture.suffix = 'world'
-        fixture.writeAnnotationProcessorTo(file("b"))
-
-        run(':c:run')
-
-        then:
-        skipped(":a:${language.compileTaskName}")
-        executedAndNotSkipped(":b:${language.compileTaskName}")
-        executedAndNotSkipped(":c:${language.compileTaskName}")
-        outputContains('hello world')
-    }
-
-    def "ignores annotation processor implementation when included in the compile classpath but annotation processing is disabled"() {
-        given:
-        Assume.assumeTrue(language == Language.JAVA)
-        settingsFile << "include 'c'"
-
-        buildFile << """
-            project(':b') {
-                dependencies {
-                    compile project(':a')
-                }
-            }
-            project(':c') {
-                dependencies {
-                    compile project(':b')
-                }
-                ${language.compileTaskName}.options.annotationProcessorPath = files()
-            }
-        """
-
-        def fixture = new HelperProcessorFixture()
-
-        fixture.writeSupportLibraryTo(file("a"))
-        fixture.writeApiTo(file("b"))
-        fixture.writeAnnotationProcessorTo(file("b"))
-
-        file("c/src/main/${language.name}/TestApp.${language.name}") << '''
-            @Helper
-            class TestApp {
-                public static void main(String[] args) {
-                }
-            }
-'''
-
-        when:
-        run(":c:${language.compileTaskName}")
-
-        then:
-        executedAndNotSkipped(":a:${language.compileTaskName}")
-        executedAndNotSkipped(":b:${language.compileTaskName}")
-        executedAndNotSkipped(":c:${language.compileTaskName}")
-
-        when:
-        run(":c:${language.compileTaskName}")
-
-        then:
-        skipped(":a:${language.compileTaskName}")
-        skipped(":b:${language.compileTaskName}")
-        skipped(":c:${language.compileTaskName}")
-
-        when:
-        // Update the library class
-        fixture.message = 'hello'
-        fixture.writeSupportLibraryTo(file("a"))
-
-        run(":c:${language.compileTaskName}")
-
-        then:
-        executedAndNotSkipped(":a:${language.compileTaskName}")
-        skipped(":b:${language.compileTaskName}")
-        skipped(":c:${language.compileTaskName}")
-
-        when:
-        // Update the processor class
-        fixture.suffix = 'world'
-        fixture.writeAnnotationProcessorTo(file("b"))
-
-        run(":c:${language.compileTaskName}")
-
-        then:
-        skipped(":a:${language.compileTaskName}")
-        executedAndNotSkipped(":b:${language.compileTaskName}")
-        skipped(":c:${language.compileTaskName}")
     }
 
     def "change to transitive super-class in different project should trigger recompilation"() {
@@ -1089,163 +820,5 @@ public class A extends B {
 
         then:
         failure.assertHasCause('Compilation failed; see the compiler error output for details.')
-    }
-
-    private String goodAstTransformation() {
-        """
-import org.codehaus.groovy.transform.*;
-import org.codehaus.groovy.ast.*;
-import org.codehaus.groovy.control.*;
-@GroovyASTTransformation(phase = CompilePhase.CANONICALIZATION)
-public class MyASTTransformation extends AbstractASTTransformation {
-    @Override
-    public void visit(ASTNode[] nodes, SourceUnit source) {
-            System.out.println("Hello from AST transformation!");
-    }
-}
-    """
-    }
-
-    private String goodAstTransformationWithABIChange() {
-        """
-import org.codehaus.groovy.transform.*;
-import org.codehaus.groovy.ast.*;
-import org.codehaus.groovy.control.*;
-@GroovyASTTransformation(phase = CompilePhase.CANONICALIZATION)
-public class MyASTTransformation extends AbstractASTTransformation {
-    @Override
-    public void visit(ASTNode[] nodes, SourceUnit source) {
-            System.out.println("Hello from AST transformation!");
-    }
-    public void foo() { }
-}
-    """
-    }
-
-    private String badAstTransformationNonABIChange() {
-        """
-import org.codehaus.groovy.transform.*;
-import org.codehaus.groovy.ast.*;
-import org.codehaus.groovy.control.*;
-@GroovyASTTransformation(phase = CompilePhase.CANONICALIZATION)
-public class MyASTTransformation extends AbstractASTTransformation {
-    @Override
-    public void visit(ASTNode[] nodes, SourceUnit source) {
-        assert false: "Bad AST transformation!"
-    }
-    public void foo() { }
-}
-    """
-    }
-
-    def "recompile with change of local ast transformation"() {
-        given:
-        Assume.assumeTrue(language == Language.GROOVY)
-        executer.beforeExecute {
-            executer.withArgument('--info')
-        }
-        buildFile << """
-            project(':b') {
-                configurations { astTransformation }
-                dependencies {
-                    astTransformation project(':a')
-                }
-                
-                tasks.withType(GroovyCompile) {
-                    compilerPluginClasspath.from(configurations.astTransformation)
-                }
-            }
-        """
-        file("a/src/main/groovy/MyAnnotation.groovy") << """
-import java.lang.annotation.*;
-@Retention(RetentionPolicy.RUNTIME)
-@Target([ElementType.TYPE])
-@org.codehaus.groovy.transform.GroovyASTTransformationClass("MyASTTransformation")
-public @interface MyAnnotation { }
-        """
-        def astTransformationSourceFile = file("a/src/main/groovy/MyASTTransformation.groovy")
-        file("b/src/main/groovy/Main.groovy") << """
-            @MyAnnotation 
-            public class Main { }
-        """
-
-        when:
-        astTransformationSourceFile << goodAstTransformation()
-
-        then:
-        succeeds ":b:compileGroovy"
-        outputContains('Hello from AST transformation!')
-        outputContains('Groovy compilation avoidance is an incubating feature')
-        executedAndNotSkipped ":a:compileGroovy"
-        executedAndNotSkipped ":b:compileGroovy"
-
-        when:
-        astTransformationSourceFile.text = goodAstTransformationWithABIChange()
-
-        then:
-        succeeds ":b:compileGroovy"
-        outputContains('Hello from AST transformation!')
-        executedAndNotSkipped ":a:compileGroovy"
-        executedAndNotSkipped ":b:compileGroovy"
-
-        when:
-        astTransformationSourceFile.text = badAstTransformationNonABIChange()
-
-        then:
-        fails ":b:compileGroovy"
-        executedAndNotSkipped ":b:compileGroovy"
-        failure.assertHasCause('Bad AST transformation!')
-    }
-
-    def "recompile with change of global ast transformation"() {
-        given:
-        Assume.assumeTrue(language == Language.GROOVY)
-        executer.beforeExecute {
-            executer.withArgument('--info')
-        }
-        buildFile << """
-            project(':b') {
-                configurations { astTransformation }
-                dependencies {
-                    astTransformation project(':a')
-                }
-                
-                tasks.withType(GroovyCompile) {
-                    compilerPluginClasspath.from(configurations.astTransformation)
-                }
-            }
-        """
-        file("a/src/main/resources/META-INF/services/org.codehaus.groovy.transform.ASTTransformation") << "MyASTTransformation"
-        def astTransformationSourceFile = file("a/src/main/groovy/MyASTTransformation.groovy")
-        file("b/src/main/groovy/Main.groovy") << """
-            public class Main { }
-        """
-
-        when:
-        astTransformationSourceFile << goodAstTransformation()
-
-        then:
-        succeeds ":b:compileGroovy"
-        outputContains('Hello from AST transformation!')
-        outputContains('Groovy compilation avoidance is an incubating feature')
-        executedAndNotSkipped ":a:compileGroovy"
-        executedAndNotSkipped ":b:compileGroovy"
-
-        when:
-        astTransformationSourceFile.text = goodAstTransformationWithABIChange()
-
-        then:
-        succeeds ":b:compileGroovy"
-        outputContains('Hello from AST transformation!')
-        executedAndNotSkipped ":a:compileGroovy"
-        executedAndNotSkipped ":b:compileGroovy"
-
-        when:
-        astTransformationSourceFile.text = badAstTransformationNonABIChange()
-
-        then:
-        fails ":b:compileGroovy"
-        executedAndNotSkipped ":b:compileGroovy"
-        failure.assertHasCause('Bad AST transformation!')
     }
 }
