+++ subprojects/plugins/src/integTest/groovy/org/gradle/java/compile/IncrementalJavaCompileAvoidanceAgainstJarIntegrationSpec.groovy
@@ -16,7 +16,8 @@
 
 package org.gradle.java.compile
 
-class IncrementalJavaCompileAvoidanceAgainstJarIntegrationSpec extends AbstractJavaGroovyCompileAvoidanceAgainstJarIntegrationSpec {
+class IncrementalJavaCompileAvoidanceAgainstJarIntegrationSpec extends AbstractJavaCompileAvoidanceIntegrationSpec implements AbstractJavaGroovyCompileAvoidanceAgainstJarIntegrationSpec {
     Language language = Language.JAVA
     boolean incremental = true
+    boolean useJar = true
 }
