+++ subprojects/plugins/src/integTest/groovy/org/gradle/java/compile/NonIncrementalJavaCompileAvoidanceAgainstJarIntegrationSpec.groovy
@@ -16,7 +16,8 @@
 
 package org.gradle.java.compile
 
-class NonIncrementalJavaCompileAvoidanceAgainstJarIntegrationSpec extends AbstractJavaGroovyCompileAvoidanceAgainstJarIntegrationSpec {
+class NonIncrementalJavaCompileAvoidanceAgainstJarIntegrationSpec extends AbstractJavaCompileAvoidanceIntegrationSpec implements AbstractJavaGroovyCompileAvoidanceAgainstJarIntegrationSpec {
     Language language = Language.JAVA
     boolean incremental = true
+    boolean useJar = true
 }
